p: replace shell script with Scheme implementation
To avoid requiring a shell and truncate the file path portably.

(And to be able to output a newline after the vm output and still be
able to pass on the vm's exit code. Although adding the newline
somehow doesn't give the intended effect (thus not included in the
patch here), and passing on the exit code portably is not done yet.)

Signed-off-by: Christian Jaeger <pflanze@gmx.ch>