Change the parser to use syntax objects.
There currently is no direct benefit (and it makes the code clunky in
places), but this will make it easier to use Racket's expander, and to
do optimizations based on identifier provenance (optimizing the
stdlib's `+', for instance).

These will become possible once we use identifiers (as opposed to
symbols) in envs.